# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/matches/flooring": {
            "get": {
                "description": "API for retrieving flooring matches based on user preferences.\nRetrieves matches based on flooring preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Retrieve flooring matches",
                "parameters": [
                    {
                        "type": "string",
                        "example": "'{\"address_lon\": 13.45, \"address_lat\": 52.5, \"services\": [\"tiles\", \"carpet\"], \"floor_size\": 120.5, \"phone_number\": \"123-456-7890\"}'",
                        "description": "QueryParams Object",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.QueryParams"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.QueryParams": {
            "type": "object",
            "required": [
                "address_lat",
                "address_lon",
                "floor_size",
                "phone_number",
                "services"
            ],
            "properties": {
                "address_lat": {
                    "type": "number"
                },
                "address_lon": {
                    "type": "number"
                },
                "floor_size": {
                    "type": "number"
                },
                "phone_number": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}